requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
boto3==1.26.94
botocore==1.29.94
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
Flask-Misaka==1.0.0
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
misaka==2.1.1
multidict==6.0.4
openai==0.27.2
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
soupsieve==2.4
tqdm==4.65.0
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.8.2
